requirements.txt
transformers
diffusers
ftfy==6.1.1
opencv-python==4.8.1.78
einops==0.7.0
pytorch-lightning==1.9.0
bitsandbytes==0.44.0
prodigyopt==1.0
lion-pytorch==0.0.6
came_pytorch==0.1.3
schedulefree==1.4
tensorboard
safetensors==0.4.4
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.25.2
imagesize==1.4.1
numpy<=2.0
rich==13.7.0
sentencepiece==0.2.0
httpx==0.24.1
accelerate
peft==0.15.2
datasets
wandb
dotenv
graphviz